move gantt to component
