+++ client/components/index.jsx
@@ -1,5 +1,6 @@
+import Gantt from "./Gantt";
 import Head from "./Head";
 import Stage from "./Stage";
 import TipCard from "./TipCard";
 
-export { Head, Stage, TipCard };
+export { Gantt, Head, Stage, TipCard };
